third_party/blink/web_tests/webexposed/storage-estimate-usage-details.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
promise_test(async function() {
    const estimate = await navigator.storage.estimate();
    assert_true('usageDetails' in estimate);
}, 'usageDetails should be exposed by navigator.storage.estimate().');
</script>
